Fixes issue #730-fix-error-msg in connection show
Change error message so it correctly shows the name that was causing the
problem.

Adds test.

We changed a couple of the other tests that use stdin because not
putting a command into the first line of the stdin causes click
to output a help message before the first command is processed.
This eliminates that extra help output from the output.

Marked one test in test_generaloptions.py failed since and filed
issue #730 on the apparent issue.