Refactor pywbemcli startup code
pywbemcli uses python files that are executed during the startup
process to define the mock of selected component as part of
the pywbemcli tests.  Generally we mock the prompt and
associated functions to that we can test pywbemcli commands that
use prompt and continue based on the return.  This involves
commands that 1. pick from a list, 2. Prompt for a confirm (Y/N)
and enter the password before the server is connected.

Whereas the old code used python files defined with
the --mock-server general option but with the first line with
as particular string to execute these files as part of the
startup of the server, the new code uses a separate environment
variable that should be considered private to pywbemcli to define
the python file to be executed as part of the pywbemcli startup.

The environment variable we use is "pywbemcli_startup_script".