Fixes issue #661-IssuesWithConnectionFileFailMsgs (#729)
* Fixes issue #661-IssuesWithConnectionFileFailMsgs

Add unit tests for connection repository to test for errors in
the YAML file being loaded.

This now outputs exceptions for all significant errors in the
yaml load including:

1. Any YAML parse error.
2. Missing key names
3. Unknown keys for server elements
4  Missing file

The exceptions attempt to point as close as possible to the
name of the file, the element causing the problem and the
exception. Note that connections file load issues are considered
fatal and terminate the interactive session.

Created new class hierarchy for exceptions in ConnectionRepository.py.

Fixes issue issue to allow setting the default back to None and adds test
for that change.

Refactor get and set default_connection_name methods.
Removed these methods since we included both a property and a method.
This required extending the getter to force the lazy load and creating a
setter method.

Move all use of click.exceiption and click.Abort() to the the
_cmd_connection.py and pywbemcli.py files so that the
ConnectionRepository.py only uses non-click exceptions.

Modified exception in pywbemcli to not abort if issue found with
default_connection_name . Do not abort of the default_connection_name
defines a non-existent connection.  Just fix it, continue and
terminate that command rather than abort the interactive mode.
All other issues with the connections file abort the interactive mode

* Fixes per PR comments.

This commit includes all the changes created in response to the issues
raised in the previous review.

* WIP

* WIP