Fixes issue #751.
Add pyline warning to all uses of set_trace() to eliminate the
warning that appears if pdbpp is used in place of pdb.  Since
pdbpp is supposed to be a direct replacement for pdb (at least in
linux environments), this seems logical. Note that this is not
a requirement since there is no required use of pdbpp but since it
does simplify debugging its use is logical.