Fixes issue #350, resetting of general options to their default in interactive mode.
This pr modifies the behavior to provide a way to reset at least some of
the general options in interactive mode by using the value "" (Empty
string) as the value when the value is input.

The general options that are modified include: --user, --password,
--keyfile, --certfile, --default-namespace, --log, --output-format,
--connection-repo.

A number of the flag general options were already covered by modifying
them to use the option that allows setting to all possible values (ex.
-v, --verbose/--noverbose)

Added tests for these resets.  Note that the log and output format are
hard to test since there is no display that shows the current value so
we are just executing the change and not testing the result.