Fixes issue # 766-Pywbemcli tests fail, assertion mof_compiler
The tests actually fail because of a badly defined qualifier declaration
in the model qualifier_filter_model.mof.  The Indication qualifier
declaration is invalid it this test (scope should be class, indication).

Fixed.

The secondary issue of an assertion error is an issue in pywbem, not
pywbemcli

NOTE: cmdshelp.rst was also modified because we must have missed that
change in earlier commit.   Build doc made the changes seen.