Fixes issue #183 define a mock model that separates the interop namespace functionality from user namespace functionality.
This change modifies the test class WbemServerMock to create the
multiple namespaces including interop and user namespace, and
inserts the model components required for each namespace into
that namespace, and  builds the instances and classes for a
basic set of profiles including the central instances of one
profile.

This produces a model with multiple profiles defined and the
central instances defined for a profile in the user nameespace.

Kept the old mock script interface for now.

Found ordering issue in server info command

The command was not ordering the namespace list if it included multiple
namespaces so that the order could be different in pywthon 2 than python
3. Modified the code to order the list for display

This PR leaves the mock script startup using the global variables
because it was getting difficult to do both without separating the
startup from the class itself into separate files.

Found a spelling error in the the help for profile commands and
fixed.