fixes issue # 810 - add option --summary to class find command
Adds the option --summary to the class find command so that only
the count of classes per namespace is displayed rather than the
complete list of classes.

It honors the -sort so that the namespaces are displayed either in
random or alphabetic order.

It is useful as part of determining how large a namespace is.

Modified the output slightly when output_format is txt. The output
format in this case was namespace:data.  Modified slightly to
include space after the :.