Fixes issue #682; add more filter options.
Adds filters for version, subclass-of, and schema that apply
to class enumerate, class find and instance count.

Adds a number of  tests for these new options and for combinations
of options depending very heavily on the class enumerate as the
test basis since class enumerate, class find, and instance count
all use the same code basis for the filter options.

Adds a table of the filters to documentation and one example. Adds
the .. table. sphinx directive to the table but I don't see any
special result of that.