Fixes issue #817, Extends the class find to display more information
See issue #817 for details of the output format.

This PR adds the code to optionally display extra information on the
line in the tree for each class and adds a test for this optional
capability.

Note that it adds a new test MOF file specifically for testing the variations
of this display.