Fixes issue #822, exception when mock version 4.0.3 used.
Creates new local functions for warnings.warn and warnings.warn_explicit
in _utils.py that install our local format functionality and removes the
patch decorator from pywbemcli.py.

NOTE: This fix communicates the state of command line warn options between
pywbemcli cli funcion and the local warn functions using global. This
is beause the path using the obj context, etc. involves many minor
changes to pass the warn flag throughout the system when it is only
required for the local pywbemcliwarn and pywbemcliwarn_explicit
functions.

However, since an import is only executed once, it deletes the local
version of the global after use forcing the reimport. An alternative
would have been to place the new functions in pywbemcli.py which is
already too big.

Adds one test for generating a warning with python 2.7. Note that there
are no warnings defined for python 3.5 and greater so that cannot
currently be tested in a functional test.