Fix issue with stdin test in instance tests
Fix issue and enable test that was marked failed in test_instance.py.
The issue description was completely removed in the test code and the
test enabled with some corrections.
The underlying issue had been fixed when issue #459 (ordring of prompt
lists) was committed some time ago.

No change to changes.rst required since this was only an update to a
test.