Fixes issue # 761- add comment field to connection definition
Adds a new field to connection definitions and a new command to the
connection group to edit data into the field.

connection edit allows editing the comment of any defined connection and
allows re-editing the text at will.

The editor  used is built into prompt-toolkit and allows multiline
editing including adding and deleting lines, moving the cursor with
arrow keys or the mouse, and saving the resulting comment.