Issue # 588 - Improve support for operation statistics
Improves the support for gathering statistics by:

1. Adding a new command group to enable/disable the statistics gathering
on the server and show the current statistics. Client enable of
statistics gathering is still in the -T general option.

The statiscs group commands include server-on, server-off, status,
reset, show,  and server-stats-show. The last is included but not implemented.
See issue #895

2. Modified the --timestats general option to allow time statistics to
be displayed for each command executed or only when requested with
the command "statistics show".

3. Add a couple sections to commands.rst including a whole new section
for the statistics command group.

4. Added tests for the new group and the changes to the --timestat
general options.

5. Added the structure for a statistics command to display the
statistics gathered in the server (server-stats-show) but did not
implement the command.  Returns exception, not implemented now.  Filed
issue for that extension.

Changes made per comments