Fixes issue#895 - implement statistics server-show
Implements the command to get statistical data from the server
and display it. This displays the data from the WBEM server
CIM_CIMOMStatisticalData class as a table showing the
number of operations and statistics for those operations (size,
server execution time and provider execution time) as a table
with data for each operation that has non-zero number of
operations.

Adds a simple test of server show-statistics command

This makes a number of changes to the tests/unit/testmock directory:

1. Renames the original wbemserver_mock.py to wbemserver_mock_class.py
and removes the script call code from this file.  This also cleans up
some of the code in this test file and simplifies the file.

2. Creates a new file wbemserver_mock.py that incorporates just the
pywbemcli script call code and import of the new
wbemserver_mock_class.py so that a user can create a separate config
dictionary to be used by the WBEMServer_mock class to build a mock
repository.

3.Add a file containing some instances of statistical data pulled from
OpenPegasus to be used in tests of the statistics server-show command.

4. Removes a file in testmock (dmtf_mof_schema_def.py that was not being
used.

The pr adds a test to test_statistics.py to execute the statistics-show
command after building a mock WBEM server and determines that at least a
report was built.