Remove all remaining TODOs from pywbemtools directory.
This includes the following TODOS:

1. association_shrub
   a. General TODO about more general width algorithm.  Not worth saving
   b. Comment about use of recreation of much of a wbemurl method.
Changed to note explaining why.

2. _command_statistics.py - complete doc item

3. _common.py
   a. Note about using curses in pick library.  This comment is way
beyond TODO. If we ever consider it we can make an issue then
   b. toposort support for case insensitive sort.  Changed to issue

4. _common_options - Comment about using lists vs single objects.
Expaneded the explanation of how we work and changed the code for the
qualifier filters to actually only use the list on the aggregated
object. Also changed documentation on the add_options method to be sure
users understand that it is locked to click decorator.

5. _display-cimobjects.py
   a. Changed TODO about table display of class to Issue # 249.
   b. Changed TODO on use of units to issue #953

6. _disiplaytree.py - Removed TODO that was effectively redundant.

7. pywbemcli.py -  Removed comment about trying to avoid deepcopy because I
don't think we can avoid the deepcopy

This removes all TODOs in the directory pywbemtools/pywbemtools

There are still about 70 TODOS between the docs (1), design (8), tools,
pywbemtools(Makefile (8), requirements.txt, etc.) ,github, and test
directories. Most are in tests