Add new pywbemcli command group subscriptions
This adds the new command group subscriptions. This command
group includes commands to add, delete, and list subscriptions,
filters, and listener destinations.

It also includes a command to request test indications from
selected servers.

Add test with add-destination that uses the wbem_serve_mock.py
model.

This pr extends the wbem_server_mock_class to include install of the
subscription user providers defined in pywbem.

This pr also  implements a change to the
pywbemcli/testmock/wbem_server_mock.py so that it
correctly/conditionally calls the correct setup for the mock script.

The wbem_server_mock.py script is used heavily in the tests of execution
of subscription commands because they use the interop namespace.

NOTE: In accord with comments, display_cimObjects has been modified
to allow creating tables for cim_object display where strings are
displayed without quotes around the strings.  This is included
as an option.

Change to cmd_subscription

Changed to simplify options and add Arguments

This includes changes to move from using ..._id and ..-name to a single
IDENTITY argument in the commands for add and remove to represent
the objects to be manipulated.

Adds new mock prompt to respond with 1 to tests

Adds the new mock prompt response to pywbemcli test files to respond to
a select request with 1 The existing prompt response files respond with
0 and 3.

Adds new test mof file

Adds the mof file mock_subscriptiontest.mof that is used in some of the
tests where filters and desination instances must be built that cause
identity conflicts.