Implement the schema subcommand from issue #444
Implements the schema subcommand for the server group. This command
outputs a table of the different CIM schemas and the maximum version
of the classes for either each namespace or the defined namespace.

It provides for a detailed report where the class counts, max version,
and whether there are experimental elements for each schema in each
namespace or just the totals for the namespace.