Add test for pywbem_mock disabled
This commit creates a new test script
simple_mock_model_disable_pull.py that starts sets
disable_pull_operations on the pywbem_mock server.  As part of that
it cleans up some of the other test scripts as follows:

1. Extends the code in the wbemserver_mock.py script to use both of
setup methods. It changes the name to wbemserver_mock_script.py so it is
clearly recognized as a script.

2. It also modifies the simple_interop_mock_script.py slightly  without
changing functionality. This script is used as the template for the
new simple_mock_model

3. Finally it modifies the import code for the WBEMServer_mock_script.py
to make it the same for all of the tests.

This pr modifies the tests in test_general_options.py to better test for
the behavior of the --use-pull general options and also adds a test
to confirm that when pull is disabled, the first open gets an exception.

Add script to test pywbem_mock with pull disabled.

Also added log to PywbemServer.__repr__