Fix issue #1023, interactive mode log control
Modified the use of --log in interactive mode to:

1. Restore either the command line value of --log if one interactive
command modifies the mode and the subsequent command does not include
--log. Thus, in the interactive mode, the --log general options only
lasts for that single command.  This may involve using the --log all=off
value to turn off the log for the subsequent command.

2. If the --log general option was not included reset the log in
WBEMConnection.

3. Added a verbose output showing interactive changes to log status
since the changes are complex.

Reorganized PywbemServer for clarity

Moved code to execute connect to its own method and the log configure
to its own method.

Added a copy method to PywbemServer to be able to use the copy()
method in WBEMConnection and its subclasses