Remove restrictions on additional general options with --name general option
We originally restricted the use of general options such as --user, etc. when
the general --name option was used.  This pr removes this restrictioni
except for --server and --mock-server. See issue # 1034.

Additionally, we were allowing general options that apply only to the
WBEMConnection to be included on the command line with the --mock-server
option.  This pr generates an exception for those cases. See issue: #1035

Adds tests for the changes