Extends capability to set default connection name, issue # 1039
1. adds --set-default to connection save command
2. adds command connection set-default that allows
   setting and clearing the default connection attribute without
   selecting or saving a connection definition
3. adds tests for the changes
4. adds documentation on the default connection name and the
   command changes.

I did this one the other night when I realized that even I had forgotten
how to use the default connection name and could not remember it.

Adding to the save is logical because that is often when you might
want to set the connection you saves to default.

Connection set-default command makes the whole idea of a default
connection  public and also allows the user to clear the default
if that is ever needed.