Add functions that pick from console to return indexes
Added new functions to pywbemtools/pywbemcli/_common.py to return the
index to the item(s)  selected by the users in addition to the current
functions that return the actual objects.  This allows picking items
from the console using more undestatandable data than items like paths
since the data displayed does not have to be the object to be selected.

Added test for empty list for the functions

Added test for None or empty list in the function and added code for the
the user that might have an issue  to cover that case.  Note that this
condition should never occur where the test was added because there
should never be a list with no connections (We delete the list when
the last connection is removed) but there could be a manually
created empty connections file.