Extend extra click support to include mutually exclusive option validation. This allows defining an option as mutually exclusive with other options in the same command
Adds a new class that allows defining mutually exclusive options in a
command so that the validation is part of the command processing and not
manually inserted into the command action function we write.

The the format for a mutually exclusive option is to add the
parameter mutually_exclusive<list of option names that are mutually
exclusive to the command>.

An optional parameter show_mutual_exclusive is boolean and if True
adds text to the help to indicate which other options are mutually
exclusive with the option.

Implement first use of this with the general options --server,
--name, and --mock-server.

Note that while click contrib option-groups does a better job in
that it can create both groups and mutually exclusive options it
is only available for python 3.6+.