Add more tests to test_subscription_cmd.py
Note that this also makes some largely editorial changes to the
subscriprion command code. Thus we converted the varaible owned to
owned_flag to clarify between owned_choice the choice option and
the boolean owned/permanent.

Removed code that had been in pywbemcli because of an issue in pywbem
that was fixed.