issue #1076, remove mininum-constraints-base.txt file.
This PR  removes the minimum-constraints-base.txt file and also makes
some other changes to the pywbemtools Makefile to bring it into line
with the pywbem Makefile.

We kept the changes down to:

1. Move the the contents of minimum-constraints-base.txt to
minimum-constraings.txt and some relalignment of lines to match the
pywbem Makefile.
2, Modify Makefile to remove references to
minimum-constraints-base.txt..
3. Editorial changes to make Makefile lines equal where there were minor
differences (ex. Makefile instead of Makefile in a lot of echo
statements.
4. Minor alignment of sections to keep the same order of targets as
pywbem Makefile.
5. Rework the install slightly