Extend subscription cmds to save subscription manager between commands
Caches the subscription manager between commands in interactive mode.
This can improve performance by keeping the existing owned subscriptions
locally rather than downloading them on each subscription command.

The cli mode still restores the owned subscriptions between commands in
the command line mode but the only difference the user will see is the
extra time to download the owned subscriptions.

We still have to enumerate the instances from the WBEM server for
permanent instances in the interactive mode.  I thought about adding a
flag to indicate that they were downloaded but decided to keep it simple
for this version.

Note that this only works as long as the server definition is not
modified in the interactive mode since the CmdSubscriptionManager object
is kept in the PywbemServer object contained in the context.  Thus, each
PywbemServer would have its own cache of subscription information.
However, since we do not cache the PywbemServer objects when an
interactive command changes the server definition, that information is
lost.

This also adds more information to the documentation on subscription
manager and cleans up some issues in the formatting of the
documentation.