Fixes #1067 - cache Subscription info between commands.
Extends pywbemcli to cache subscription information between commands in
interactive mode.  While the permanent subscription information is also
cached, it is not used since pywbem enumerates subscriptions each time a
request for permanent instances is executed.

Add location of instance of class CmdSubscriptionManager instance in PywbemServer
instance and recovery when a command is called.  The instance is in
PywbemServer object so that there is a set of cached subscription data
for each wbem server created.

Since we do not cache the PywbemServer objects if the server is changed
in interactive mode, the subscription cache only exists where the server
definition is not changed.

This also adds documentation and corrects some documentation grammar and
formatting.

No significant to the functionality of the subscription command group

Corrected one error from an earlier commit:

We were defaulting to the default-namespace for
SourceNamespaces value if not was supplied on the command line.  It now
passes None which is consistent with the specification.