Bypass for OpenPegasus issue  defined in issue #1097.
Fixes the issue with incomplete instance name in reference properties
by changing to use the path in place of the instance to acquire
reference properties.

Also refactors tests for subscription list*** so there is a test
for each subscription object (list, list-destinations, list-filters,
list-subscriptions) rather than all of the tests in one single
test.

No change to changes.rst because this is  part of new v 1.0 code.