fix issue #1118 exception possible exception when --use-pull no used
This case passes a value of None for general option pull_max_cnt and the
general option use_pull is No.  Pywbem tests for the existence of a
value (not None) before it tests for whether it is to use pull or not so this case
fails because it the value is None since no value was set.

Simple modification to always pass the resolved numeric value which is
never None.  Solved by extending PywbemServer to set the default
value in the pull_max_cnt if the input is None

Added tests against mock for the use of a WBEM operation with
all combinations of the --use-pull operation in the mock environment.

However, we should add these tests in the end-end test environment
also.  Tested manually against container and it works.