Change output format for instance names table in display_cimobject
The original output formatting for

instance enumerate --no

was useless.  It showed 3 4 columns host, ns, class, keybindings

where the keybindings was a single string.

Modified it to show"

host, namespace, class, one column per key which allows the user to
compare keys between instances.

This does not do anything for associations where the reference
properties are keys, however.

In the case of enumerates of instances starting from a point above
the hierarchy (ex. CIM_ManagedElement) it produces a table for each
set of key definitions (i.e. all the instances with the same keys
in the same table.   This means producing multiple tables.
Note also, that the elements in each table are only linked by

1. They are instances of some class in the hierarchy.
2. They have the same keynames.

They may be instances that are not strongly related (i.e. from
completely different parts of the hiearchy.

Fixes the current tests for instance enumerate ... --no