Modify the instance table view processor to improve the table display in several ways including:
1. Modify to_wbem_url_folded to fold at all major separators:

Originally it folded only for keys when the length was greater than the
overall string output and folded all keys.

Modified to fold at end of host name, namespace name, classname, and for
the full definition or name/value component of each key to improve
control of the size of the result.

It does not fold either the key name or the key value if they exceed the
max_len

2. Modify the cell width calculation to allow tables wider than the
terminal width by setting a minimum table column width if the total
number of columns (i.e. properties to display)  would make the table display
exceed the terminal width.  The alternate would have been tables with
very narrow and largely unreadable tables.

3. Modified to fold property names in the instance table view if they
exceed the computed column width. The code before did not account for
the fact that property names are single words.

Fixed one test for enumerate-names-only in test_instances

Change the test to reflect the correct format consistent with the
modified output in issue #1131

Modified a number of the table output displays to match the formatting
changes.

Make one minor change to the pywbemtools _mof_str in the cim formatter
to add maxline to the values shown in the assert for endless loop. This
will help if it ever occurs again and since I do not know exactly what
the minimum maxline is that causes the problem this at least helps.

Added to the documentation to better define the table output mechanisms.