Refactor documentation
1. added option list for all commands

2. Added the command usage line for all commands.  This does not
matter for commands without arguments but is important for any
that have arguments. Tried to mark all arguments with the double back
quote so that they can be identified in the text.

3. Moved the definition of the class result filters such as
--assocation, etc. to a common section in the features section and
referenced that from locations where they are used rather than repeating
all of these options for each command in which they are used.

4. Cleaned up a number of poor descriptions and editorial errors.