Fix issue where tests fail if keys not sorted
In order to create consistent table output for test of instances names
as table, the keys must be sorted.  Otherwise the column orders are not
consistent across different test targets, etc.  This simply sorts the
CIMInstanceName keys before the headers and rows are created for the
table.

Add tests for the changes where it caused test failures

Fixed test issue whee we were not completing all tests in one unit test.