Expand tab-completion and its documentation.
Completes the documentation of what tab-completion and auto-suggesion
are, where they exist and how to activate tab-completion for the
command line for the bash, zsh, and fish shells.

Adds option value tab completion for --name, --mock-server,
--conection-file options, --output-format, --keyfile, --certfile,
etc.

A tab-complete function was written for the options and arguments
that look up date locally (--name general option and <subject>
help argument. This option is accessed through the cls=<func name>
attribute of the click.Option/click.Argument decorator.  However,
to support the options that want to find
files or directories the only change was to modify the option
type from str to click.Path.  Further, choice now automatically
allows tab-completion so the only change was to the output-format
to include the "" as a legitimate choice.

Adds the help activate, help tab-completion subjects