Implement multiple namespaces for commands commands
The ability to process multiple command options for the following
commands is implementedi including both the case that returns
the full objects, the option to return only names, and the
summary option:

* class enumerate
* class get
* class associators
* class references
* instance enumerate
* instance get
* instance associators
* instance references

The --namespace option has been modified for each of these commands to
allow multiple namespaces.  This allows multiple namespaces in two
forms, -n root/cimv2,root/cimv3 or -n root/cimv2 -n root/cimv3.

The results displayed in display_cim_objects(...)  has been modified to
show the namespace as part of the response if multiple namespaces are
requested.  Otherwise the existing display format is maintained.

NOTE: There were a number of changes to display_cim_objects() to unify
the displays and cover some edge cases that were not covered before in
addition to the changes to accept the multi-namespace input.

For table output, the results of multiple namespaces are aggregated into
a single table. For single object independent output (xml, mof, etc.)
the output format was extended for each format to insure that the
namespace name is attached to each object. ex. The #pragma namespace
is added before each class/instance display if multiple namespaces
were requested but not if a single namespace was requested. For XML
a comment with the namespace name was added with each XML object. Where
the namespace was part of the object display, nothing was added.

Added tests to test_instance_cmds.py and test_class_cmds.py for the
case of multiple namespaces with a new test file that provides
3 namespaces (interop, root/cimv3, and root/cimv3) where the two
cim.. namespaces are almost identical.

Adds a new test script

Adds a new test script simple-three-ns-mock-script.py that includes 3
namespaces (interop, root/cimv2, and root/cimv3) where the root... ones
are almost the same except for some instance differences.