Fix issue using instance get with multiple namespace (-n) option
The "instance get <classname>.?" command with multiple namespaces (-n
root/cimv2,root/cimv3) caused an exception.  However, this led to
realizing that we were only displaying a single instance rather than
the instances from each namespace.
Corrected those issues.

I did not add to changes since the core changes for the multiple
namespaces were in this development version and this was a code
error in those changes.