Add option to sort outputs of multi-namespaces by classname
The default output order of results from the commands that allow multiple
namespaces  is namespace, object id, etc.  This allows the user to modify the
output order to display the results ordered by object id, namespace so
that objects that are similar can be easily visually compared.

This is one of the requirements of issue #1184 but was not included in
the earlier commits for this issue to reduce the number of changes in
each pr.

This is almost the final change to cover the functional requirements
of issue 1184

It affects all of the commands that process multiple objects including
class/instance/qualifier enumerate/get/associators/references.

This adds a new boolean option "--object-order" to each of the requests
defined above  that changes the order of display of cim objects for
the object formats and the table formats.

The order of namespaces whether it is the first or second sort item
is kept as the order in which the command was defined.

This pr also fixes an issue where in some cases, the cim objects being
displayed were not ordered. They are now always ordered in alphabetic
order within their grouping.

This pr also adds a number of tests to confirming that the
--object-order works.

NOTE: This pr reorganizes the display_cim_objects functions significatly
as part of the change, to  clean up the code.