Fix issue #1184 - Multiple namespace commands with errors
Originally pywbemcli always returned an exception for any error on any
namespace when multiple namespaces were requested for instance/class get,
enumerate, references, associations.

Issue #1184 requested that results returned be displayed along with warning
about errors with any of the server requests for multiple namespace
processing that failed  unless it was with all the requests or interpreted
as a error beyond the scope of a particular request.

This pr modifies the returns for the instance, class, qualifier  get, enumerate,
references, and associators table and mof views to return a warning for
each request where the request ended in exception related to existence of
the target item and then return all items found for those commands.

The original behavior of returns or exception is kept if only a single
namespace is specified or the default namespace is used.  If the user specifies
multiple namespaces pywbemcli  will request processing continue after
selected CIM errors exceptions and output the set of errors in a form
compatible with the output expected (table, etc.)

This is done for exceptions like NOT_FOUND, INVALID_CLASS,
INVALID_PARAMTER, and INVALID_NAMESPACE. All other exceptions are
treated as complete failure of the pywbemcli command.

Add new common_cmd_options module. Moved a number of functions that are
common to multiple command-groups to a separate new module.