Fix issue in instance shrub
Fix issue in instance shrub where the ResultRole element of the shrub
was shown as a CIMClassname object and retained in the internal
structure as that object resulting in duplication of target instances
possible when inter-namespace associators were defined.

Refactored the build_shrub method to simplify and make more readable.

Added more tests for the shrub command.