Issue # 1193, Fix issue with general option --pull-max-cnt
Fixes an issue that was causing this option to not be saved
correctly.  While the issue assumes that the option value is
not stored in the PywbemServer object but in the ContextObj, in fact,
the code is in place but a flag was misset so that the existence of
this option in an interactive command was not updating the PywbemServer.
Added a test to confirm the persistence of this and other
options.