Add object ordering to multi-namespace requests to allow ordering by object name (#1189)
* Fix issue #1184 - Multiple namespace commands with errors

Originally pywbemcli always returned an exception for any error on any
namespace when multiple namespaces were requested for instance/class get,
enumerate, references, associations.

Issue #1184 requested that results returned be displayed along with warning
about errors with any of the server requests for multiple namespace
processing that failed  unless it was with all the requests or interpreted
as a error beyond the scope of a particular request.

This pr modifies the returns for the instance, class, qualifier  get, enumerate,
references, and associators table and mof views to return a warning for
each request where the request ended in exception related to existence of
the target item and then return all items found for those commands.

The original behavior of returns or exception is kept if only a single
namespace is specified or the default namespace is used.  If the user specifies
multiple namespaces pywbemcli  will request processing continue after
selected CIM errors exceptions and output the set of errors in a form
compatible with the output expected (table, etc.)

This is done for exceptions like NOT_FOUND, INVALID_CLASS,
INVALID_PARAMTER, and INVALID_NAMESPACE. All other exceptions are
treated as complete failure of the pywbemcli command.

Add new common_cmd_options module. Moved a number of functions that are
common to multiple command-groups to a separate new module.

* Add option to sort outputs of multi-namespaces by classname

The default output order of results from the commands that allow multiple
namespaces  is namespace, object id, etc.  This allows the user to modify the
output order to display the results ordered by object id, namespace so
that objects that are similar can be easily visually compared.

This is one of the requirements of issue #1184 but was not included in
the earlier commits for this issue to reduce the number of changes in
each pr.

This is almost the final change to cover the functional requirements
of issue 1184

It affects all of the commands that process multiple objects including
class/instance/qualifier enumerate/get/associators/references.

This adds a new boolean option "--object-order" to each of the requests
defined above  that changes the order of display of cim objects for
the object formats and the table formats.

The order of namespaces whether it is the first or second sort item
is kept as the order in which the command was defined.

This pr also fixes an issue where in some cases, the cim objects being
displayed were not ordered. They are now always ordered in alphabetic
order within their grouping.

This pr also adds a number of tests to confirming that the
--object-order works.

NOTE: This pr reorganizes the display_cim_objects functions significatly
as part of the change, to  clean up the code.

Co-authored-by: kschopmeyer <k.schopmeyerwork@gmail.com>