Fix issue #1145, adding classname to instance table output
Adds a classname column to instance, associators, references table output
when the display is more than one class.

Also adds flag if the --deep-inheritance flag is set since that
can affect what shows up in table output.

Note that when only a single class is represented in the table
output, the classname already is in the output (i.e. instance name)

This required modifying a lot of tests so there are many modifications
to test_instance_cmd.py.