Fix issue in test module wbemserver_moc_class.py
There is a case in this code where an instance is created of an
association but the path for one of the references includes the host
element in CIMInstanceName object. Documented in isue #1203 and fixed
in pywbem issue #2920

This was invalid but only with pywbem 1.5.0 is that validated. Not
setting the host element allows passing tests with pywbem 1.5.0