Clean up documentation.
A number of changes to pywbemcli documentation including":

1. Move the new general options table to more logical location in the
documentation and some changes to text in the table. Issue # 1163

2. Add more references between pywbemcli and pywbem docs including
references back to the MOF compiler and other components defined
in the pywbem documentation.

3. Extend documentation of the --leaf-classes option.

4, Wording cleanup in a number of places.

5. Extension to the section on mock support to explain mock scripts
and to reference back to pywbem documentation for more information.
This heavily changes the mock support section. issue #1209