Fix issue where instance shrub fails tests with tabulate 0.9.0
Fixed issue where tabulate update to 0.9.0 causes failure of tests.

Issue was a not backward compatible change by tabulate and issue with
out code where we were sending CIMClassName object to tabulate.

Further it became obvious that we we were displaying more than we need
to with the associator class entry in the table which was sending the
whole CIMClassname to display as classname.

Changed to only send classname string.

This required changing a number of tests that were displaying the whole
URL for the CIMClassName element representing the associator class.